combine fetch and evaluation into a unified identification service #6
