+++ src/services/index.ts
@@ -1,6 +1,4 @@
 import { FillService } from './fill-service';
-import { EvaluationService } from './evaluation-service';
-import { FetchService } from './fetch-service';
-import { preparationService } from './preparation-service';
+import { IdentificationService } from './identification-service';
 
-export { FillService, EvaluationService, FetchService, preparationService };
+export { FillService, IdentificationService };
